requirements.txt
flask
flask-sqlalchemy
flask-login
python-dotenv
flask-limiter
itsdangerous
resend
pyotp
gunicorn
psycopg2-binary
werkzeug
email-validator
cryptography
Flask-WTF
Flask-Talisman